only allow one process to execute Discogs scrapers simultaneously otherwise the api will lock us out when multiple Discogs scrapers are queued
the problem most likely is that the Celery rate limiting only assures that tasks are not **started** with a higher frequency than specified, the Discogs api does its accounting for rate limiting probably also after the request is served. so if requests run longer than 1 second (which is most of them) new tasks could have been started that will also make requests which will look to the api like we try to make more than 1 req/s and will get us blocked.