Link to MusicBrainz using HTTPS.
Some of the ones in descgen/tests.py should probably remain http:// to
test that pasting in a HTTP link will still work and spit out a HTTPS
one, but the workflow of the tests didn't seem obvious, so I skipped
this for now. New tests can be made or the tests can have these changes
reverted at a later point.